refactor(agent-commerce): A8 — extract Hero/MoversBoard/ActivityPulse (P2 precondition for redesign)
Pure refactor. Splits 3 inline sections from src/app/agent-commerce/page.tsx
(~2293 lines) into co-located components under src/components/agent-commerce/
so the upcoming visible-redesign PRs (hero band, 2-col body) can land as
readable diffs.

New files:
  - Sparkline.tsx                       — shared SVG helper
  - AgentCommerceHero.tsx               — page-head + verdict ribbon + KPI band
  - AgentCommerceMoversBoard.tsx        — composite movers board (with per-row meta)
  - AgentCommerceActivityPulse.tsx      — 12-week spark + protocol pulse

Page reduction: 2293 → ~1817 lines (-476 net). MiniBoard stays inline (used
only in the leaderboards section); the BrandStar / Metric / MetricGrid /
pluralize imports it brought in are removed because they were only used by
the now-extracted sections.

Behavior bytes-identical: sec-head numbers, class names, grid templates,
inline styles, and conditional protocol metas all preserved. Pure
presentation — every value passes in via props.

Verified: npm run typecheck, npm run lint:guards, npm run test:hooks all
pass on this branch.
